+++ futurefinity/server.py
@@ -21,7 +21,8 @@ FutureFinity Web Application, which can parse http request, initialize
 right RequestHandler and make response to client.
 """
 
-from futurefinity.utils import *
+from futurefinity.utils import ensure_str, ensure_bytes
+from futurefinity.protocol import HTTPHeaders, HTTPRequest, HTTPError
 
 import futurefinity
 
@@ -55,17 +56,14 @@ class HTTPServer(asyncio.Protocol):
         self._loop = loop
         self.app = app
         self.enable_h2 = enable_h2
+
         self.transport = None
         self._keep_alive_handler = None
-        self.data = b""
-        self._crlf_mark = None
+
+        self._request_parser = None
+        self._request_finished = False
+
         self.http_version = 10
-        self.initial = None
-        self._body_parsed = False
-        self.method = ""
-        self.content_type = ""
-        self.content_length = 0
-        self.parsed_body = None
 
         self._request_handlers = {}
 
@@ -93,15 +91,9 @@ class HTTPServer(asyncio.Protocol):
         """
         self._request_handlers = {}
         self.set_keep_alive_handler()
-        self.data = b""
-        self._crlf_mark = None
-        self.initial = None
-        self._request_handler = None
-        self._body_parsed = False
-        self.method = ""
-        self.content_type = ""
-        self.content_length = 0
-        self.parsed_body = None
+
+        self._request_parser = None
+        self._request_finished = False
 
     def connection_made(self, transport: asyncio.BaseTransport):
         """
@@ -175,91 +167,42 @@ class HTTPServer(asyncio.Protocol):
             if self.http_version == 20:
                 return  # HTTP/2 will be implemented later.
             else:
-                self.data_received_http_v1(data)
+                self.http_v1_data_received(data)
         except Exception as e:
             if self.app.settings.get("debug", False):
                 traceback.print_exc()
             self.handle_request_error(e)
 
-    def data_received_http_v1(self, data: bytes):
-        """
-        Try to parse received data as HTTP/1.x request.
-
-        This function should not be used directly, data_received() function
-        will pass it to the right http version.
-        """
-        if self._body_parsed:
+    def http_v1_data_received(self, data: bytes):
+        if self._request_finished:
             return
 
-        self.data += data
+        if self._request_parser is None:
+            self._request_parser = HTTPRequest()
 
-        if self._crlf_mark is None:
-            self._crlf_mark = decide_http_v1_mark(
-                self.data[:MAX_HEADER_LENGTH + 1])
+        self._request_finished = self._request_parser.parse_http_v1_request(
+            data)
+        self.http_version = self._request_parser.http_version
 
-            if self._crlf_mark is None:
-                return  # Request Not Completed, Wait.
+        if self._request_finished:
+            self._request_handlers[0] = asyncio.ensure_future(
+                self.handle_request(self._request_parser))
 
-            self.initial, self.data = parse_http_v1_initial(
-                self.data, use_crlf_mark=self._crlf_mark)
-
-            self.http_version = self.initial["http_version"]
-            self.method = self.initial["parsed_headers"][":method"]
-            if self.method in BODY_EXPECTED_METHODS:
-                self.content_type = self.initial[
-                    "parsed_headers"].get_first("content-type")
-                self.content_length = int(
-                    self.initial["parsed_headers"].get_first("content-length"))
-
-        if self.method in BODY_EXPECTED_METHODS:
-            self.parse_body_http_v1()
-            if not self._body_parsed:
-                return  # Request Not Completed, wait.
-        else:
-            self._body_parsed = True
-
-        if len(self._request_handlers.keys()) != 0:
-            raise HTTPError(500)
-            # HTTP/1.x should have only one RequestHandler at the same time.
-
-        self._request_handlers[0] = asyncio.ensure_future(
-            self.handle_request(self.initial, self.parsed_body))
-
-    def parse_body_http_v1(self):
-        """
-        Try to Parse Data as HTTP/1.x Request Body.
-        """
-        if len(self.data) < self.content_length:
-            return  # Request Not Completed, wait.
-        self.parsed_body = parse_http_v1_body(
-            data=self.data,
-            content_type=self.content_type,
-            content_length=self.content_length
-        )
-        self._body_parsed = True
-
-    async def handle_request(self, initial: dict,
-                             parsed_body: cgi.FieldStorage):
+    async def handle_request(self, request: HTTPRequest):
         """
         Handle an HTTP Request to Right RequestHandler.
         """
-        matched_obj = self.app.find_handler(initial["parsed_path"])
+        matched_obj = self.app.find_handler(request.path)
         request_handler = matched_obj.pop("__handler__")(
             app=self.app,
             server=self,
-            method=initial["parsed_headers"][":method"],
-            path=initial["parsed_path"],
-            matched_path=matched_obj,
-            queries=initial["parsed_queries"],
-            http_version=self.http_version,
-            request_headers=initial["parsed_headers"],
-            request_cookies=initial["parsed_cookies"],
-            request_body=parsed_body,
+            request=request,
             make_response=self.make_response
         )
         await request_handler.handle(**matched_obj)
 
-    def make_response(self, status_code: int, response_headers: HTTPHeaders,
+    def make_response(self, status_code: int,
+                      response_headers: HTTPHeaders,
                       response_body: bytes):
         """
         Make http response to client.
@@ -270,9 +213,11 @@ class HTTPServer(asyncio.Protocol):
             self.make_http_v1_response(status_code, response_headers,
                                        response_body)
 
-    def make_http_v1_response(self, status_code: int,
-                              response_headers: HTTPHeaders,
-                              response_body: bytes):
+    def make_http_v1_response(
+        self,
+        status_code: int,
+        response_headers: HTTPHeaders,
+        response_body: bytes):
 
         """
         Make HTTP/1.x response to client.